Tests

+++ NotesWindowsFormsApp/TagDatabaseRepository.cs
@@ -1,14 +1,19 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 
 namespace NotesWindowsFormsApp
 {
-    class TagDatabaseRepository : ITagRepository
+    public class TagDatabaseRepository : ITagRepository
     {
-        internal TaskContext context;
-        public void Add(Tag tag)
+        private readonly TaskContext context;
+        public TagDatabaseRepository(TaskContext context)
         {
-            context.Tags.Add(tag);
+            this.context = context;
+        }
+        public void Add(Tag tag)
+        {                      
+            context.Tags.Add(tag);                      
             context.SaveChanges();
         }
         public Tag FindByText(string text)
@@ -24,10 +29,14 @@ namespace NotesWindowsFormsApp
             }
             return listOfTags;
         }
-        public void Delete(Tag tagToDelete)
-        {            
-            context.Tags.Remove(tagToDelete);
-            context.SaveChanges();
+        public void Delete(string textOfTagToDelete)
+        {
+            var tagToDelete = FindByText(textOfTagToDelete);
+            if (tagToDelete != null)
+            {
+                context.Tags.Remove(tagToDelete);
+                context.SaveChanges();
+            }
         }
 
     }
